# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guest-place",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/about": {
      "get": {
        "tags": [
          "About"
        ],
        "operationId": "get_about",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AboutResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AboutResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AboutResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "About"
        ],
        "operationId": "update_about",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAboutDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "About page updated"
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    },
    "/api/v1/analytics/engagement": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_engagement_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Engagement stats by entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_EngagementStatsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/funnel": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_funnel_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversion funnel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_FunnelStepDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/referrals": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_referral_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Referral sources",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_ReferralDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/session": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "operationId": "start_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SessionResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/summary": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_summary_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AnalyticsSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/track": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "operationId": "track_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackEventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Event tracked"
          }
        }
      }
    },
    "/api/v1/analytics/traffic": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "operationId": "get_traffic_stats",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to look back",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Traffic stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TrafficStatsDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/logout/other": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "logout_other_devices",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshTokenDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Другие сессии завершены"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Ошибка базы данных"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshTokenDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Успешное обновление токена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Refresh-токен недействителен или истек",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthResponseDTO"
                }
              }
            }
          },
          "500": {
            "description": "Ошибка базы данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/register/key": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "check_register_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckEmailCodeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Код действителен"
          },
          "400": {
            "description": "Некорректный код"
          },
          "404": {
            "description": "Код не найден или истёк"
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
    "/api/v1/block-types": {
      "get": {
        "tags": [
          "Pages"
        ],
        "operationId": "get_block_types",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_BlockType"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Pages"
        ],
        "operationId": "create_block_type",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlockTypeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BlockType"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/block-types/sync": {
      "post": {
        "tags": [
          "Pages"
        ],
        "operationId": "sync_block_types",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncBlockTypesDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_BlockType"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/block-types/{id}": {
      "delete": {
        "tags": [
          "Pages"
        ],
        "operationId": "delete_block_type",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Block Type ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Value"
                }
              }
            }
          },
          "404": {
            "description": "Block type not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Pages"
        ],
        "operationId": "update_block_type",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Block Type ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlockTypeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BlockType"
                }
              }
            }
          },
          "404": {
            "description": "Block type not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/content/entries": {
      "post": {
        "tags": [
          "Content"
        ],
        "operationId": "create_entry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentEntryDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ContentEntry"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/content/entries/{id}/versions": {
      "get": {
        "tags": [
          "Content"
        ],
        "operationId": "get_entry_versions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Entry ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_ContentEntryVersion"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_Vec_EngagementStatsDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "entity_id",
                "clicks",
                "views"
              ],
              "properties": {
                "clicks": {
                  "type": "integer",
                  "format": "int64"
                },
                "entity_id": {
                  "type": "string"
                },
                "views": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "errors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "messages": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CheckEmailCodeDTO": {
        "type": "object",
        "required": [
          "key"
        ],
        "properties": {
          "key": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "gp_apikey"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}